add cog integration with python3.8
