+++ cog.yaml
@@ -1,0 +1,60 @@
+# Configuration for Cog ⚙️
+# Reference: https://github.com/replicate/cog/blob/main/docs/yaml.md
+
+build:
+  # cuda: "11.0.3"
+  # set to true if your model requires a GPU
+  gpu: true
+
+  # a list of ubuntu apt packages to install
+  system_packages:
+    - "sox"
+    - "libsndfile1"
+    - "ffmpeg"
+
+  # python version in the form '3.8' or '3.8.12'
+  python_version: "3.8"
+
+  # a list of packages in the format <package-name>==<version>
+  python_packages:
+    - "numpy==1.22.0"
+    - "scipy==1.7.0"
+    - "tensorflow==2.8.0"
+    - "tensorboard==2.8.0"
+    - "psola==0.0.1"
+    - "wget==3.2"
+    - "torch==1.8.1"
+    - "torchaudio==0.8.1"
+    - "torchtext==0.9.1"
+    - "librosa==0.9.2"
+    - "torch-stft==0.1.4"
+    - "kaldiio==2.17.2"
+    - "pydub==0.25.1"
+    - "pyannote.audio==1.1.2"
+    - "frozendict==2.0.3"
+    - "inflect==6.0.0"
+    - "unidecode==1.2.0"
+    - "nltk==3.7"
+    - "g2p-en==2.1.0"
+    - "editdistance==0.5.3"
+    - "pesq==0.0.2"
+    - "pystoi==0.3.3"
+    - "crepe==0.0.12"
+    - "resampy==0.2.2"
+    - "ffmpeg-python==0.2.0"
+    - "tqdm==4.64.1"
+    - "gdown==4.5.1"
+    - "ipywidgets==7.6.3"
+    - "torchcrepe==0.0.15"
+    - "taming-transformers-rom1504==0.0.6"
+    - "einops==0.3.2"
+    - "tensorflow-hub==0.12.0"
+    - "git+https://github.com/SortAnon/NeMo.git"
+  
+  # commands run after the environment is setup
+  # run:
+    # - "echo env is ready!"
+    # - "echo another command if needed"
+
+# predict.py defines how predictions are run on your model
+predict: "predict.py:Predictor"
