+++ talknet_offline.py
@@ -1,17 +1,17 @@
-if __name__ == "__main__":
-    import os
-    from controllable_talknet import *
+# if __name__ == "__main__":
+#     import os
+#     from controllable_talknet import *
 
-    if os.path.exists("/talknet/is_docker"):
-        app.run_server(
-            host="0.0.0.0",
-            mode="external",
-            debug=False,
-            threaded=True,
-        )
-    else:
-        app.run_server(
-            mode="external",
-            debug=False,
-            threaded=True,
-        )
+#     if os.path.exists("/talknet/is_docker"):
+#         app.run_server(
+#             host="0.0.0.0",
+#             mode="external",
+#             debug=False,
+#             threaded=True,
+#         )
+#     else:
+#         app.run_server(
+#             mode="external",
+#             debug=False,
+#             threaded=True,
+#         )
